+++ modules/examples/eb-sf-batch/tests/test_stack.py
@@ -1,2 +1,47 @@
-def test_placeholder() -> None:
-    return None
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+import os
+import sys
+
+import aws_cdk as cdk
+import pytest
+from aws_cdk.assertions import Template
+
+
+@pytest.fixture(scope="function")
+def stack_defaults():
+    os.environ["ADDF_PROJECT_NAME"] = "test-project"
+    os.environ["ADDF_DEPLOYMENT_NAME"] = "test-deployment"
+    os.environ["ADDF_MODULE_NAME"] = "test-module"
+    os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
+    os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
+
+    if "stack" in sys.modules:
+        del sys.modules["stack"]
+
+
+def test_synthesize_stack(stack_defaults):
+    import stack
+
+    app = cdk.App()
+    dep_name = "test-deployment"
+    mod_name = "test-module"
+
+    event_driven_batch = stack.EventDrivenBatch(
+        scope=app,
+        id=f"addf-{dep_name}-{mod_name}",
+        deployment_name=dep_name,
+        module_name=mod_name,
+        fargate_job_queue_arn="arn:foobar",
+        ecr_repo_name="myrepo",
+        vcpus=2,
+        memory_limit_mib=2048,
+        env=cdk.Environment(
+            account=os.environ["CDK_DEFAULT_ACCOUNT"],
+            region=os.environ["CDK_DEFAULT_REGION"],
+        ),
+    )
+
+    template = Template.from_stack(event_driven_batch)
+    template.resource_count_is("AWS::Batch::JobDefinition", 1)